Fixed unreachable code from NodeJS fixed, also making same interface for NodeJS and Bower
