+++ lib/eid/exceptions.js
@@ -17,7 +17,7 @@
 (function(exports) {
   'use strict';
 
-  var Eid = require('../eid');
+  var Eid = require('./eid');
   var commonExConstructor = function(eid, message) {
     if (!(eid instanceof Eid)) {
       eid = new Eid(eid.toString());
