+++ lib/eid/preconditions.js
@@ -17,7 +17,7 @@
 (function(module) {
 'use strict';
 
-var Eid = require('../eid');
+var Eid = require('./eid');
 var EidRuntimeException = require('./exceptions').EidRuntimeException;
 var EidNullPointerException = require('./exceptions').EidNullPointerException;
 var EidIllegalArgumentException = require('./exceptions').EidIllegalArgumentException;
